refactor(T9292): retire legacy executor.ts (cleoLlmCallInner) — migrate to ConcreteExecutor (W3c)
- Delete executor.ts and migration-snapshot-gemini.test.ts (referenced deleted backends)
- Extract CompletionResult/ProviderBackend/etc. shims to legacy-types.ts (TODO T9298 W5)
- Provide cleoLlmCallInner stub in shim-executor.ts (always throws — ProviderBackend removed)
- Update api.ts, tool-loop.ts, history-adapters.ts, structured-output.ts, llm-layer.test.ts
- Export ConcreteExecutor, DefaultLlmExecutorFactory, getLlmExecutor from index.ts
